feature/GE-630: Added new app, moved model, added to settings, TODO: Move forms, views and urls
