+++ authentication_service/user_migration/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class UserMigrationConfig(AppConfig):
+    name = 'user_migration'
